add point module, update node
